[Behat][Admin] Minor clean up in promotion pages
