+++ src/Sylius/Behat/Page/Admin/Promotion/UpdatePage.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Sylius\Behat\Page\Admin\Promotion;
 
 use Behat\Mink\Element\NodeElement;
-use Behat\Mink\Exception\ElementNotFoundException;
 use Sylius\Behat\Behaviour\ChecksCodeImmutability;
 use Sylius\Behat\Behaviour\CountsChannelBasedErrors;
 use Sylius\Behat\Behaviour\NamesIt;
@@ -25,16 +24,6 @@ class UpdatePage extends BaseUpdatePage implements UpdatePageInterface
     use ChecksCodeImmutability;
     use CountsChannelBasedErrors;
     use NamesIt;
-
-    public function setPriority(?int $priority): void
-    {
-        $this->getDocument()->fillField('Priority', $priority);
-    }
-
-    public function getPriority(): int
-    {
-        return (int) $this->getElement('priority')->getValue();
-    }
 
     public function checkChannelsState(string $channelName): bool
     {
@@ -131,21 +120,6 @@ class UpdatePage extends BaseUpdatePage implements UpdatePageInterface
         return $this->countChannelErrors($this->getElement('rules'), $channelCode);
     }
 
-    /**
-     * @throws ElementNotFoundException
-     */
-    public function getValidationMessageForTranslation(string $element, string $localeCode): string
-    {
-        $foundElement = $this->getElement($element, ['%localeCode%' => $localeCode])->getParent();
-
-        $validationMessage = $foundElement->find('css', '.sylius-validation-error');
-        if (null === $validationMessage) {
-            throw new ElementNotFoundException($this->getSession(), 'Validation message', 'css', '.sylius-validation-error');
-        }
-
-        return $validationMessage->getText();
-    }
-
     protected function getCodeElement(): NodeElement
     {
         return $this->getElement('code');
@@ -163,7 +137,6 @@ class UpdatePage extends BaseUpdatePage implements UpdatePageInterface
             'ends_at_time' => '#sylius_promotion_endsAt_time',
             'exclusive' => '#sylius_promotion_exclusive',
             'coupon_based' => '#sylius_promotion_couponBased',
-            'label' => '#sylius_promotion_translations_%localeCode%_label',
             'name' => '#sylius_promotion_name',
             'order_percentage_action_field' => '[id^="sylius_promotion_actions_"][id$="_configuration_percentage"]',
             'priority' => '#sylius_promotion_priority',
